requirements.txt
# Smart Assistant Requirements
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# GUI
PyQt6==6.6.1

# Email & IMAP
imaplib2==3.6

# Database & Storage
sqlite3
faiss-cpu==1.7.4

# NLP & LLM
openai==1.3.7
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2

# Scheduler
APScheduler==3.10.4

# Security
keyring==24.3.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
asyncio

# GUI & Exe Building
PyInstaller==6.3.0
Pillow==10.1.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
